third_party/WebKit/LayoutTests/media/media-source-append-multiple-expected.txt

EVENT(loadstart)
EVENT(loadedmetadata)
loadedMetadataA
EVENT(progress)
EVENT(suspend)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
EXPECTED (audio.currentSrc == sourceA.src == 'true') OK
END OF TEST
